backend/agent.py
from schemas import MeetingContext, WhiteboardData, WhiteboardElement
from google.cloud import speech, texttospeech
import uuid

class ScrumAgent:
    def __init__(self):
        # Initialize Google clients, memory, etc.
        self.speech_client = speech.SpeechClient()
        self.tts_client = texttospeech.TextToSpeechClient()
        # ... other state

    def get_initial_context(self):
        # Return a MeetingContext with predefined participants, agenda, etc.
        return MeetingContext(
            id=str(uuid.uuid4()),
            status="not_started",
            participants=[
                {"id": "1", "name": "Alice", "role": "Frontend"},
                {"id": "2", "name": "Bob", "role": "Backend"},
                {"id": "3", "name": "Carol", "role": "Product"}
            ],
            current_speaker=None,
            agenda=["Opening", "Updates", "Blockers", "Summary"],
            notes=[],
            whiteboard=WhiteboardData(elements=[]),
            memory={}
        )

    async def start_meeting(self, context: MeetingContext):
        context.status = "active"
        context.current_speaker = context.participants[0].name
        # Add initial whiteboard elements, etc.

    async def next_participant(self, context: MeetingContext):
        # Move to next participant, update context
        idx = next((i for i, p in enumerate(context.participants) if p.name == context.current_speaker), 0)
        next_idx = (idx + 1) % len(context.participants)
        context.current_speaker = context.participants[next_idx].name

    async def process_speech(self, audio: bytes, context: MeetingContext):
        # Use Google STT to transcribe
        audio_obj = speech.RecognitionAudio(content=audio)
        config = speech.RecognitionConfig(
            encoding=speech.RecognitionConfig.AudioEncoding.LINEAR16,
            sample_rate_hertz=16000,
            language_code="en-US"
        )
        response = self.speech_client.recognize(config=config, audio=audio_obj)
        transcript = response.results[0].alternatives[0].transcript if response.results else ""
        # Parse transcript, update context/whiteboard
        # Example: If "blocker" in transcript, add sticky note
        if "blocker" in transcript.lower():
            context.whiteboard.elements.append(
                WhiteboardElement(
                    id=f"blocker-{context.current_speaker}-{uuid.uuid4()}",
                    type="sticky",
                    x=100, y=100,
                    text=f"{context.current_speaker}: {transcript}",
                    color="red"
                )
            )
        # Save transcript to memory, etc.
        context.memory[context.current_speaker] = transcript
        return transcript

    # Add TTS methods as needed
